refactor: extract show function to a file
